Restructure @dnd-mapp/data project
