+++ packages/data/src/index.ts
@@ -1,1 +1,1 @@
-export * from './lib/models';
+export * from './lib';
