+++ packages/data/src/lib/api/index.ts
@@ -1,0 +1,3 @@
+export * from './ability.model';
+export * from './race.model';
+export * from './skill.model';
