+++ packages/data/src/lib/api/skill.model.ts
@@ -1,5 +1,5 @@
+import { NameableEntityModel } from '../models';
 import { Ability } from './ability.model';
-import { NameableEntityModel } from './entity.model';
 
 export interface Skill extends NameableEntityModel {
     ability: Ability;
