+++ packages/data/src/lib/authentication/index.ts
@@ -1,0 +1,2 @@
+export * from './user-role.model';
+export * from './user.model';
