+++ packages/data/src/lib/models/index.ts
@@ -1,6 +1,1 @@
-export * from './ability.model';
 export * from './entity.model';
-export * from './race.model';
-export * from './skill.model';
-export * from './user-role.model';
-export * from './user.model';
